Fix errors and most warnings when compiling with gcc (GCC) 5.3.1 20151207 (Red Hat 5.3.1-2)

+++ src/gay/dtoa.c
@@ -227,7 +227,7 @@ extern void *MALLOC(size_t);
 #ifndef PRIVATE_MEM
 #define PRIVATE_MEM 2304
 #endif
-#define PRIVATE_mem ((PRIVATE_MEM+sizeof(double)-1)/sizeof(double))
+#define PRIVATE_mem ((unsigned)((PRIVATE_MEM+sizeof(double)-1)/sizeof(double)))
 static double private_mem[PRIVATE_mem], *pmem_next = private_mem;
 #endif
 
@@ -3574,7 +3574,7 @@ rv_alloc(int i)
 
 	j = sizeof(ULong);
 	for(k = 0;
-		sizeof(Bigint) - sizeof(ULong) - sizeof(int) + j <= i;
+		(int)(sizeof(Bigint) - sizeof(ULong) - sizeof(int) + j) <= i;
 		j <<= 1)
 			k++;
 	r = (int*)Balloc(k);
